Move map code to MapController
