+++ dsnake4-phaser/src/components/Game/Data/Generics.ts
@@ -1,4 +1,6 @@
 export type Direction = 'Up' | 'Right' | 'Left' | 'Down';
 
+export type MapCellType = 'Wall' | 'Pickup' | 'Stairs' | 'SnakePart' | 'Void';
+
 export const CELLS_X: number = 80;
 export const CELLS_Y: number = 60;
