+++ dsnake4-phaser/src/components/Game/Data/MapController.ts
@@ -1,80 +1,116 @@
-import { CELLS_Y, CELLS_X } from './Generics';
-import { MapCellType, MapVector, MapCell, MapElement } from './MapElements';
+import { MapVector, MapCell } from './MapElements';
+import { GameObjects, Scene } from 'phaser';
+import { BodyPart, Snake } from './Snake';
+import { KeyBindings } from './KeyBindings';
+import { JustDown } from '../imports';
+import { Map } from './Map';
 
 export class MapController {
-    /* 
-    This class should calculate the map from all children elements.
-    Do note that if children overlap, order of entry matters (we don't check overlap).
-    
-    Also note that calculating the final map should be done once (or at least not a lot).
-    */
-    private childElements: MapElement[];
-    public Map2D!: MapCell[][];
+    private scene: Scene;
+    private map: Map;
+    private snake: Snake;
 
-    constructor() {
-        this.childElements = [];
-        this.constructMap();
+    cellHeight: number;
+    cellWidth: number;
+    inputKeys!: KeyBindings;
+
+    private defaultTextStyle = {
+        fontSize: this.cellHeight?.toString(),
+        fontStyle: 'normal',
+        fontFamily: 'Consolas',
+        color: "#42b983",
+        backgroundColor: 'rgba(0,0,0,0)'
+    };
+
+    constructor(scene: Scene, cellWidth: number, cellHeight: number) {
+        this.scene = scene;
+
+        this.cellHeight = cellHeight;
+        this.cellWidth = cellWidth;
+        this.map = new Map();
+        this.snake = new Snake(15, 16, 3, 'Right');
+
+        console.log("MapController constructed with cell size", this.cellHeight, this.cellWidth);
     }
 
-    public constructMap() {
+    setup() {
+        this.inputKeys = this.scene.input.keyboard.addKeys('W,UP,S,DOWN,A,LEFT,D,RIGHT') as KeyBindings;
+        this.renderSnake();
+    }
+
+    checkInput() {
+        if (JustDown(this.inputKeys.W) || JustDown(this.inputKeys.UP)) {
+            this.snake.rotateUp();
+        } else if (JustDown(this.inputKeys.A) || JustDown(this.inputKeys.LEFT)) {
+            this.snake.rotateLeft();
+        } else if (JustDown(this.inputKeys.S) || JustDown(this.inputKeys.DOWN)) {
+            this.snake.rotateDown();
+        } else if (JustDown(this.inputKeys.D) || JustDown(this.inputKeys.RIGHT)) {
+            this.snake.rotateRight();
+        }
+    }
+
+    checkSnakeCollision() {
+        console.log("Snake checked at ", this.snake.x, this.snake.y);
+        return this.map.checkCollision(this.snake.x, this.snake.y) == 'Wall';
+    }
+
+    timedUpdate() {
+        this.snake.moveSnake();
+        this.renderSnake();
+    }
+
+    render() {
+        console.log("Controller start");
+        this.constructMap();
+
+        this.renderMap();
+        this.renderSnake();
+    }
+
+    private constructMap() {
         const offsetX = 30;
         const offsetY = 30;
-        this.appendElement(new MapVector(new MapCell(offsetX, offsetY, 'Wall'), 3, 'Up'));
-        this.appendElement(new MapVector(new MapCell(offsetX, offsetY, 'Wall'), 3, 'Down'));
-        this.appendElement(new MapVector(new MapCell(offsetX, offsetY, 'Wall'), 3, 'Left'));
-        this.appendElement(new MapVector(new MapCell(offsetX, offsetY, 'Wall'), 3, 'Right'));
-        // this.appendElement(new MapVector(new MapCell(offsetX, offsetY, 'Stairs'), 30, 'Right'));
-        // this.appendElement(new MapVector(new MapCell(10, 1, 'Wall'), 30, 'Right'));
-        // this.appendElement(new MapVector(new MapCell(1, 30, 'Wall'), 30, 'Right'));
+
+        // Load elements into the map
+        this.map
+            .appendElement(new MapVector(new MapCell(offsetX, offsetY, 'Wall'), 3, 'Up'))
+            .appendElement(new MapVector(new MapCell(offsetX, offsetY, 'Wall'), 3, 'Down'))
+            .appendElement(new MapVector(new MapCell(offsetX, offsetY, 'Wall'), 3, 'Left'))
+            .appendElement(new MapVector(new MapCell(offsetX, offsetY, 'Wall'), 3, 'Right'));
+
+        // Perform processing to 2D-array
+        this.map.flattenMap();
     }
 
-    public checkCollision(x: number, y: number): MapCellType {
-        if (x <= 0 || x >= CELLS_X) {
-            console.log('hit x wall', x);
-            return 'Wall';
-        }
-        if (y <= 0 || y >= CELLS_Y) {
-            console.log('hit y wall', x);
-            return 'Wall';
-        }
-        if (this.Map2D[x] == null || this.Map2D[x][y] == null) {
-            return 'Void';
-        }
-        return this.Map2D[x][y].type;
+    private renderMap() {
+        this.map.Map2D
+            .forEach(row => row
+                .forEach(cell => {
+                    this.scene.add.rectangle(
+                        cell.x * this.cellWidth - this.cellWidth / 2,
+                        cell.y * this.cellHeight - this.cellHeight / 2,
+                        this.cellWidth - 2, this.cellHeight - 2,
+                        0xEEEEEE);
+                }));
     }
 
-    public getMapCell(x: number, y: number) {
-        return this.Map2D[x][y];
-    }
-
-    // This function performs the pre-processing to get an 2D-Array of cells in `this.Map2D`
-    public flattenMap() {
-        this.Map2D = [];
-        if (this.childElements != null) {
-            this.childElements.forEach(elem => {
-                elem.cells.forEach(elemCell => {
-                    if (this.Map2D[elemCell.x] == null) {
-                        this.Map2D[elemCell.x] = [];
-                    }
-                    this.Map2D[elemCell.x][elemCell.y] = elemCell.clone();
-                })
+    private renderSnake() {
+        if (this.snake?.bodyParts != null) {
+            this.snake.bodyParts.forEach(part => {
+                this.renderSnakePart(part);
             });
         }
     }
 
-    // public loadMapFromImage(imageMap: any) {
-    //     throw new Error("The loadMapFromImage function has not been implemented yet.");
-    // }
-
-    public clearChildren() {
-        this.childElements = [];
-    }
-
-    public clear2DMap() {
-        this.Map2D = [];
-    }
-
-    public appendElement(element: MapElement) {
-        this.childElements.push(element);
+    private renderSnakePart(part: BodyPart) {
+        const pixelX = (part.x - 1) * this.cellWidth + 1;
+        const pixelY = (part.y - 1) * this.cellHeight - 2;
+        if (part.gameObject == null) {
+            part.gameObject = this.scene.add.text(pixelX, pixelY, part.toCharacter(), this.defaultTextStyle);
+        }
+        else {
+            part.gameObject.setPosition(pixelX, pixelY);
+        }
     }
 }
