+++ dsnake4-phaser/src/components/Game/Data/MapElements.ts
@@ -1,6 +1,4 @@
-import { Direction, CELLS_X, CELLS_Y } from './Generics';
-
-export type MapCellType = 'Wall' | 'Pickup' | 'Stairs' | 'Void';
+import { Direction, CELLS_X, CELLS_Y, MapCellType } from './Generics';
 
 export class MapCell {
     /*
